+++ src/main/java/english/chatbot/application/RegisterUserService.java
@@ -1,6 +1,6 @@
 package english.chatbot.application;
 
-import english.chatbot.application.entity.User;
+import english.chatbot.application.domain.entity.User;
 import english.chatbot.application.port.in.RegisterUserUseCase;
 import english.chatbot.application.port.out.RegisterUserPort;
 import lombok.RequiredArgsConstructor;
