+++ src/main/java/english/chatbot/application/UpdateUserService.java
@@ -1,6 +1,6 @@
 package english.chatbot.application;
 
-import english.chatbot.application.entity.User;
+import english.chatbot.application.domain.entity.User;
 import english.chatbot.application.port.in.UpdateUserUseCase;
 import english.chatbot.application.port.out.FindUserPort;
 import jakarta.transaction.Transactional;
