+++ src/main/java/english/chatbot/application/port/in/FindUserUseCase.java
@@ -1,6 +1,6 @@
 package english.chatbot.application.port.in;
 
-import english.chatbot.application.entity.User;
+import english.chatbot.application.domain.entity.User;
 
 public interface FindUserUseCase {
